+++ test/examples/single-axis-label-scaling.ts
@@ -1,12 +1,37 @@
-import { AnchorType, createLayer, DataProvider, LabelInstance, LabelLayer, LayerInitializer, ScaleType } from '../../src';
+import { AnchorType, ChartCamera, createLayer, DataProvider, LabelInstance, LabelLayer, LayerInitializer, ReferenceCamera, ScaleType } from '../../src';
 import { BaseExample } from './base-example';
 
-export class LabelAnchorsAndScales extends BaseExample {
+export class SingleAxisLabelScaling extends BaseExample {
+  isYAxis: boolean = true;
+
+  constructor(yAxis: boolean = true) {
+    super();
+    this.isYAxis = yAxis;
+  }
+
+  makeCamera(defaultCamera: ChartCamera): ChartCamera {
+    if (this.isYAxis) {
+      return new ReferenceCamera({
+        base: defaultCamera,
+        offsetFilter: (offset: [number, number, number]) => [0, offset[1], 0],
+        scaleFilter: (scale: [number, number, number]) => [1, scale[1], 1],
+      });
+    }
+
+    else {
+      return new ReferenceCamera({
+        base: defaultCamera,
+        offsetFilter: (offset: [number, number, number]) => [offset[0], 0, 0],
+        scaleFilter: (scale: [number, number, number]) => [scale[0], 1, 1],
+      });
+    }
+  }
+
   makeLayer(scene: string, atlas: string, provider: DataProvider<LabelInstance>): LayerInitializer {
     return createLayer(LabelLayer, {
       atlas,
       data: provider,
-      key: 'label-anchors-and-scales',
+      key: this.isYAxis ? 'vertical-label-scaling' : 'horizontal-label-scaling',
       scene,
     });
   }
@@ -30,7 +55,7 @@ export class LabelAnchorsAndScales extends BaseExample {
       rasterization: {
         scale: 1.0,
       },
-      scaling: ScaleType.NEVER,
+      scaling: ScaleType.BOUND_MAX,
       text: 'Anchored MiddleLeft:',
       x: 20,
       y: count++ * 20,