+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 1.1.0
+
+* `(ADDED)` A base layer property 'picking' is now available for setting PickType.ALL, PickType.SINGLE and PickType.NONE
+* `(ADDED)` Layers can implement quad tree mouse interactions now for the PickType.ALL mode
+* `(ADDED)` Base circle layer has quad tree mouse interactions available now
+
 ## 1.0.1
 
 * `(FIXED)` auto release notes errored on first release